clean up AST types
- address later feedback on #719
- fix syntax highlighting for TS templates
- use `ReadonlyArray<T>` in return types instead of `Array<T>`.
- use `undefined` instead of `null` for optional fields, as it is more idiomatic.
