+++ crates/codegen/parser/generator/src/lib.rs
@@ -1,9 +1,11 @@
 mod ast_model;
-mod code_generator;
 mod parser_definition;
 mod precedence_parser_definition;
+mod rust_generator;
 mod scanner_definition;
 mod trie;
+mod typescript_generator;
 
 pub use ast_model::AstModel;
-pub use code_generator::CodeGenerator;
+pub use rust_generator::RustGenerator;
+pub use typescript_generator::TypeScriptGenerator;
