+++ documentation/public/user-guide/npm-package/how-to-parse-a-file/index.md
@@ -63,11 +63,11 @@ Let's do that:
 --8<-- "crates/solidity/outputs/npm/tests/src/doc-examples/reconstruct-source.ts:step-2-assertion"
 ```
 
-### Example 2: Reading Contracts using Cursors
+### Example 2: Listing contract names using Cursors
 
 The `Cursor` type provides procedural-style functions that allow you to navigate the source in a step-by-step manner. In addition to `goToNext`, we can go to the parent, first child, next sibling, etc., as well as nodes with a given kind.
 
-Let's start with this piece of solidity source code:
+Let's start with this piece of Solidity source code:
 
 ```{ .ts }
 --8<-- "crates/solidity/outputs/npm/tests/src/doc-examples/list-contract-names.ts:step-1-imports"
@@ -81,11 +81,11 @@ To list the top-level contracts and their names, we need to visit the `ContractD
 --8<-- "crates/solidity/outputs/npm/tests/src/doc-examples/list-contract-names.ts:step-3-cursor"
 ```
 
-### Example 3: Reading Contracts using AST Types
+### Example 3: Listing contract names using AST Types
 
 AST types are a set of TypeScript classes that provide a typed view of the untyped CST nodes.
 You can convert any untyped CST node to its corresponding AST type using their constructors.
-AST types are immutable, and are constructed on the fly, as they are used/accessed for the first time.
+AST types are immutable. Additionally, their fields are constructed lazily as they are accessed for the first time.
 
 Let's try to perform the same task using AST types:
